+++ tests/test_builder.py
@@ -8,8 +8,13 @@ from pathlib import Path
 import numpy as np
 import scipy as sp
 from openfermion import QubitOperator, count_qubits, get_sparse_operator
+from pyscf.fci import FCI
+from pyscf.gto import Mole
+from pyscf.scf import RHF, UHF
+from pytest import raises
 
 from nbed.driver import NbedDriver
+from nbed.exceptions import HamiltonianBuilderError
 from nbed.ham_builder import HamiltonianBuilder
 
 logger = getLogger(__name__)
@@ -20,103 +25,132 @@ args = {
     "geometry": str(mol_filepath),
     "n_active_atoms": 1,
     "basis": "STO-3G",
-    "xc_functional": "b3lyp",
-    "projector": "mu",
-    "localization": "spade",
+    # "xc_functional": "b3lyp",
+    # "projector": "mu",
+    # "localization": "spade",
     "convergence": 1e-8,
-    "savefile": None,
-    "run_ccsd_emb": False,
-    "run_fci_emb": False,
+    # "run_ccsd_emb": False,
+    # "run_fci_emb": False,
 }
+
+mol_args = {
+    "atom": str(mol_filepath),
+    "n_active_atoms": 1,
+    "basis": "STO-3G",
+    "unit": "angstrom",
+}
+
+mol = Mole(**mol_args, charge=0, spin=0).build()
+restricted_scf = RHF(mol)
+restricted_scf.kernel()
+
+unrestricted_scf = UHF(mol)
+unrestricted_scf.kernel()
 
 
 def test_restricted() -> None:
     """
     Use the full system to check that output hamiltonian diagonalises to fci value for a restricted calculation.
     """
-    restric_driver = NbedDriver(
-        geometry=args["geometry"],
-        n_active_atoms=args["n_active_atoms"],
-        basis=args["basis"],
-        xc_functional=args["xc_functional"],
-        projector=args["projector"],
-        localization=args["localization"],
-        convergence=args["convergence"],
-        savefile=args["savefile"],
-        run_ccsd_emb=args["run_ccsd_emb"],
-        run_fci_emb=args["run_fci_emb"],
-        force_unrestricted=False,
-    )
 
-    fci = restric_driver._global_fci.e_tot - restric_driver._global_ks.energy_nuc()
-    logger.info(f"FCI energy of unrestricted driver test: {fci}")
+    e_fci = FCI(restricted_scf).kernel()[0] - restricted_scf.energy_nuc()
 
-    builder = HamiltonianBuilder(restric_driver._global_ks, 0, "jordan_wigner")
+    logger.info(f"FCI energy of unrestricted driver test: {e_fci}")
+
+    builder = HamiltonianBuilder(restricted_scf, 0, "jordan_wigner")
     ham = builder.build()
     diag, _ = sp.sparse.linalg.eigsh(get_sparse_operator(ham), k=1, which="SA")
     logger.info(f"Ground state via diagonalisation: {diag}")
-    assert np.isclose(fci, diag)
+    assert np.isclose(e_fci, diag)
 
 
-def test_force_unrestricted() -> None:
+rbuilder = HamiltonianBuilder(restricted_scf, 0, "jordan_wigner")
+ubuilder = HamiltonianBuilder(unrestricted_scf, 0, "jordan_wigner")
+
+
+def test_qubit_number_reduction() -> None:
     """
-    Use the full system to check that output hamiltonian diagonalises to fci value for an unrestricted calculation.
+    Check that the qubit reduction is working as expected.
     """
-    unrestric_driver = NbedDriver(
-        geometry=args["geometry"],
-        charge=0,
-        spin=0,
-        n_active_atoms=args["n_active_atoms"],
-        basis=args["basis"],
-        xc_functional=args["xc_functional"],
-        projector=args["projector"],
-        localization=args["localization"],
-        convergence=args["convergence"],
-        savefile=args["savefile"],
-        run_ccsd_emb=args["run_ccsd_emb"],
-        run_fci_emb=args["run_fci_emb"],
-        force_unrestricted=True,
-    )
 
-    fci = unrestric_driver._global_fci.e_tot - unrestric_driver._global_ks.energy_nuc()
-    logger.info(f"FCI energy of unrestricted driver test: {fci}")
+    # We're still constructing qubit hamiltonians that double the size for restricted systems!
 
-    builder = HamiltonianBuilder(unrestric_driver._global_ks, 0, "jordan_wigner")
-    ham = builder.build()
-    diag, _ = sp.sparse.linalg.eigsh(get_sparse_operator(ham), k=1, which="SA")
-    logger.info(f"Ground state via diagonalisation: {diag}")
-    assert np.isclose(fci, diag)
+    rham = rbuilder.build()
+    assert count_qubits(rham) == 14
+    uham = ubuilder.build()
+    assert count_qubits(uham) == 14
+
+
+def test_qubit_reduction() -> None:
+
+    rham = rbuilder.build(n_qubits=-1)
+    assert count_qubits(rham) == 12
+    uham = ubuilder.build(n_qubits=-1)
+    assert count_qubits(uham) == 12
+
+
+def test_qubit_specification() -> None:
+    rham = rbuilder.build(n_qubits=12)
+    assert count_qubits(rham) == 12
+    uham = ubuilder.build(n_qubits=12)
+    assert count_qubits(uham) == 12
+
+
+def test_active_space_reduction() -> None:
+    rham = rbuilder.build(core_indices=[], active_indices=[0, 1, 2, 3, 4, 5])
+    assert count_qubits(rham) == 12
+    uham = ubuilder.build(core_indices=[], active_indices=[0, 1, 2, 3, 4, 5])
+    assert count_qubits(uham) == 12
+
+
+def test_frozen_core_validation() -> None:
+    """
+    Test the the appropriate error is raised for invalid frozen core indices.
+    """
+    rbuilder = HamiltonianBuilder(restricted_scf, 0, "jordan_wigner")
+
+    with raises(HamiltonianBuilderError, match="Core indices must be 1D array."):
+        rbuilder.build(core_indices=[[0, 1], [0, 1]], active_indices=[0, 1, 2, 3, 4, 5])
+
+    with raises(HamiltonianBuilderError, match="Active indices must be 1D array."):
+        rbuilder.build(core_indices=[0], active_indices=[[1, 2], [1, 2]])
+
+    with raises(
+        HamiltonianBuilderError, match="Core and active indices must not overlap."
+    ):
+        rbuilder.build(
+            core_indices=[0, 1, 2, 3, 4, 5], active_indices=[0, 1, 2, 3, 4, 5]
+        )
+
+    with raises(
+        HamiltonianBuilderError,
+        match="Number of core and active indices must not exceed number of orbitals.",
+    ):
+        rbuilder.build(
+            core_indices=[0, 1, 2, 3, 4, 5], active_indices=[6, 7, 8, 9, 10, 11]
+        )
+
+
+mol = Mole(**mol_args, charge=1, spin=1).build()
+unrestricted_scf = UHF(mol)
+unrestricted_scf.kernel()
 
 
 def test_unrestricted() -> None:
     """
     Check the output hamiltonian diagonalises to fci value for an unrestricted calculation with spin and charge.
     """
-    unrestric_driver = NbedDriver(
-        geometry=args["geometry"],
-        charge=1,
-        spin=1,
-        n_active_atoms=args["n_active_atoms"],
-        basis=args["basis"],
-        xc_functional=args["xc_functional"],
-        projector=args["projector"],
-        localization=args["localization"],
-        convergence=args["convergence"],
-        savefile=args["savefile"],
-        run_ccsd_emb=args["run_ccsd_emb"],
-        run_fci_emb=args["run_fci_emb"],
-        force_unrestricted=True,
-    )
+    e_fci = FCI(unrestricted_scf).kernel()[0] - unrestricted_scf.energy_nuc()
 
-    fci = unrestric_driver._global_fci.e_tot - unrestric_driver._global_ks.energy_nuc()
-    logger.info(f"FCI energy of unrestricted driver test: {fci}")
+    logger.info(f"FCI energy of unrestricted driver test: {e_fci}")
 
-    builder = HamiltonianBuilder(unrestric_driver._global_ks, 0, "jordan_wigner")
+    builder = HamiltonianBuilder(unrestricted_scf, 0, "jordan_wigner")
     ham = builder.build()
     diag, _ = sp.sparse.linalg.eigsh(get_sparse_operator(ham), k=2, which="SA")
 
+    print(diag)
     # Ground state for this charge is 2nd eigenstate
     diag = diag[1]
 
     logger.info(f"Ground state via diagonalisation: {diag}")
-    assert np.isclose(fci, diag)
+    assert np.isclose(e_fci, diag)