+++ CHANGELOG.md
@@ -21,8 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Location of explainer notebooks moved up a level to `docs/notebooks`
 - Update of dependencies
 - Python dependency updated to `>=3.8, <3.11` to pull in symmer.
-- Localization notebook updated to include concentric localization.
-- `n_qubits` input to `HamiltonianBuilder.build` now allows negative values for reduction of total size.
+- `HamBuilder.build()` now accepts MO indices for active space reduction. This overwrites `n_qubits` if used.
+### Removed
+- `savefile` argument removed from driver, as it is not used.
 
 
 ## [0.0.1] - 2022-03-07